Update to newest

+++ auto h-encore/Form1.cs
@@ -13,9 +13,11 @@ using System.IO.Compression;
 using System.Net.Http;
 using Microsoft.VisualBasic.FileIO;
 using System.Diagnostics;
+using System.Security.Cryptography;
 
 namespace auto_h_encore {
     public partial class Form1 : Form {
+
 
         public Form1() {
             InitializeComponent();
@@ -33,34 +35,96 @@ namespace auto_h_encore {
         }
 
         private void generateDirectories(string AID) {
-            info("生成工作目录...");
-            if (Directory.Exists(Reference.path_data)) Directory.Delete(Reference.path_data, true);
+            //TODO: Needs code cleanup pretty bad...
+            if (cbxDelete.Checked) {
+                info("删除旧文件...");
+                if (Directory.Exists(Reference.path_data)) Directory.Delete(Reference.path_data, true);
+                for (int i = 0; i < 4; i++) {
+                    if (!FileSystem.FileExists(Global.fileOverrides[i])) Global.fileOverrides[i] = "";
+                }
+                
+            } else {
+
+                string path = "";
+                string cleanName;
+                string md5;
+
+                for (int id = 0; id < 4; id++) {
+
+                    switch (id) {
+                        case 0:
+                            path = Reference.path_downloads + "hencore.zip";
+                            break;
+                        case 1:
+                            path = Reference.path_downloads + "pkg2zip.zip";
+                            break;
+                        case 2:
+                            path = Reference.path_downloads + "psvimgtools.zip";
+                            break;
+                        case 3:
+                            path = Reference.path_downloads + "bittersmile.pkg";
+                            break;
+                    }
+
+                    cleanName = path.Replace('/', '\\').Split('\\').Last();
+
+                    if (Global.fileOverrides[id] != null && Global.fileOverrides[id] != "") {
+                        info("File import for file " + cleanName + " valid.");
+                        continue;
+                    }
+
+                    if (FileSystem.FileExists(path)) {
+                        md5 = Utility.MD5Checksum(path);
+                        if (Reference.hashes[id] == md5) {
+                            info("文件 " + cleanName + " 已下载并且有效, 不再重新下载");
+                            Global.fileOverrides[id] = path;
+                        } else {
+                            info("文件 " + cleanName + " 已下载但 hash 不匹配, 将要重新下载.");
+                        }
+                    } else {
+                        info("文件 " + cleanName + " 未下载或者导入, 将开始下载.");
+                    }
+                }
+
+                if (Directory.Exists(Reference.path_hencore)) Directory.Delete(Reference.path_hencore, true);
+                if (Directory.Exists(Reference.path_psvimgtools)) Directory.Delete(Reference.path_psvimgtools, true);
+            }
+
             if (Directory.Exists(txtQCMA.Text + "\\APP\\" + AID + "\\PCSG90096\\")) {
-                if (MessageBox.Show("你必须删除 QCMA 目录内已存在的 bittersmile 备份. 删除么?", "警告", MessageBoxButtons.YesNo) == DialogResult.Yes) {
-                    Directory.Delete(txtQCMA.Text + "\\APP\\" + txtAID.Text + "\\PCSG90096\\", true);
+                if (MessageBox.Show("你必须删除 QCMA 目录内已存在的 bittersmile 备份. 如果你想保留它，请立刻手动转移删除么?", "警告", MessageBoxButtons.YesNo) == DialogResult.Yes) {
+                    FileSystem.DeleteDirectory(txtQCMA.Text + "\\APP\\" + txtAID.Text + "\\PCSG90096\\", DeleteDirectoryOption.DeleteAllContents);
                 } else {
                     throw new IOException("目录已存在");
                 }
             }
+
+            info("生成工作目录...");
+            if (FileSystem.FileExists(Reference.fpath_pkg2zip)) FileSystem.DeleteFile(Reference.fpath_pkg2zip);
             Directory.CreateDirectory(Reference.path_data);
             Directory.CreateDirectory(Reference.path_hencore);
             Directory.CreateDirectory(Reference.path_psvimgtools);
-            Directory.CreateDirectory(Reference.path_pkg2zip);
             Directory.CreateDirectory(Reference.path_downloads);
             incrementProgress();
         }
 
         private void downloadFiles() {
-            Utility.DownloadFile(this, true, Reference.url_hencore, Reference.path_downloads + "hencore.zip");
-            Utility.ExtractFile(this, true, Reference.path_downloads + "hencore.zip", Reference.path_hencore);
-            
-            Utility.DownloadFile(this, true, Reference.url_psvimgtools, Reference.path_downloads + "psvimgtools.zip");
-            Utility.ExtractFile(this, true, Reference.path_downloads + "psvimgtools.zip", Reference.path_psvimgtools);
-            
-            Utility.DownloadFile(this, true, Reference.url_pkg2zip, Reference.path_downloads + "pkg2zip.zip");
-            Utility.ExtractFile(this, true, Reference.path_downloads + "pkg2zip.zip", Reference.path_pkg2zip);
-            
-            Utility.DownloadFile(this, true, Reference.url_bittersmile, Reference.path_downloads + "bittersmile.pkg");
+
+            for (int id = 0; id < 4; id++) {
+                string cleanName = Reference.raws[id].Replace('/', '\\').Split('\\').Last();
+                if (Global.fileOverrides[id] != null && Global.fileOverrides[id] != "") {
+                    if (Global.fileOverrides[id] == Reference.raws[id]) info("文件 " + cleanName + " 在正确的位置, 跳过");
+                    else {
+                        info("导入 " + cleanName);
+                        FileSystem.CopyFile(Global.fileOverrides[id], Reference.raws[id], true);
+                        info("      完成!");
+                    }
+                } else {
+                    Utility.DownloadFile(this, Reference.downloads[id], Reference.raws[id]);
+                }
+                incrementProgress();
+
+                if (id != 3) Utility.ExtractFile(this, true, Reference.raws[id], Reference.paths[id]);
+            }
         }
 
         private void PackageHencore(string encKey) {
@@ -93,7 +157,6 @@ namespace auto_h_encore {
         private void lblHowToAID_LinkClicked(object sender, LinkLabelLinkClickedEventArgs e) {
             FormAID frmAid = new FormAID();
             frmAid.ShowDialog();
-            
         }
 
         private void txtAID_TextChanged(object sender, EventArgs e) {
@@ -105,22 +168,23 @@ namespace auto_h_encore {
 
             //run code on new thread to keep UI responsive
             Task.Factory.StartNew(new Action(() => {
-
+                
                 try {
                     generateDirectories(txtAID.Text);
                     downloadFiles();
                 } catch (Exception ex) {
-                    MessageBox.Show(ex.Message);
+                    //MessageBox.Show(ex.Message);
                     toggleControls(true);
                     return;
                 }
                 
+
                 try {
                     info("正在利用 pkg2zip 解包 bittersmile 试玩版...");
                     ProcessStartInfo psi = new ProcessStartInfo();
-                    psi.WorkingDirectory = Reference.path_pkg2zip;
-                    psi.Arguments = "-x \"" + Reference.path_downloads + "bittersmile.pkg\"";
-                    psi.FileName = Reference.path_pkg2zip + "pkg2zip.exe";
+                    psi.WorkingDirectory = Reference.path_downloads;
+                    psi.Arguments = "-x bittersmile.pkg";
+                    psi.FileName = Reference.fpath_pkg2zip;
                     Process process = Process.Start(psi);
                     process.WaitForExit();
                     info("      完成!");
@@ -130,10 +194,22 @@ namespace auto_h_encore {
                     toggleControls(true);
                     return;
                 }
-                
+
+                if (cbxTrim.Checked) {
+                    info("删减 bitter smile 试玩版中...");
+                    string path = Reference.path_downloads + "app\\PCSG90096\\";
+                    FileSystem.DeleteFile(path + "resource\\movie\\Opening.mp4");
+                    foreach (string k in FileSystem.GetFiles(path + "resource\\sound\\bgm\\")) FileSystem.DeleteFile(k);
+                    FileSystem.DeleteDirectory(path + "resource\\sound\\voice\\01\\", DeleteDirectoryOption.DeleteAllContents);
+                    FileSystem.DeleteDirectory(path + "resource\\sound\\se\\", DeleteDirectoryOption.DeleteAllContents);
+                    FileSystem.DeleteDirectory(path + "resource\\image\\bg\\", DeleteDirectoryOption.DeleteAllContents);
+                    FileSystem.DeleteDirectory(path + "resource\\image\\tachie\\", DeleteDirectoryOption.DeleteAllContents);
+                    info("      Done!");
+                }
+
                 try {
-                    foreach (string k in Directory.EnumerateFiles(Reference.path_pkg2zip + "app\\PCSG90096\\")) {
-                        info("移动 " + k.Split('\\').Last() + " 到 h-encore 工作...");
+                    foreach (string k in FileSystem.GetFiles(Reference.path_downloads + "app\\PCSG90096\\")) {
+                        info("移动 " + k.Split('\\').Last() + " 到 h-encore 工作目录...");
                         FileSystem.MoveFile(k, Reference.path_hencore + "\\h-encore\\app\\ux0_temp_game_PCSG90096_app_PCSG90096\\" + k.Split('\\').Last());
                     }
                 } catch (DirectoryNotFoundException ex) {
@@ -151,7 +227,7 @@ namespace auto_h_encore {
                 }
 
                 try {
-                    foreach (string k in Directory.EnumerateDirectories(Reference.path_pkg2zip + "app\\PCSG90096\\")) {
+                    foreach (string k in FileSystem.GetDirectories(Reference.path_downloads + "app\\PCSG90096\\")) {
                         info("移动 " + k.Split('\\').Last() + " 到 h-encore 工作目录...");
                         FileSystem.MoveDirectory(k, Reference.path_hencore + "\\h-encore\\app\\ux0_temp_game_PCSG90096_app_PCSG90096\\" + k.Split('\\').Last());
                     }
@@ -193,7 +269,7 @@ namespace auto_h_encore {
                 string encKey;
 
                 try {
-                    info("利用 AID 获取 CMA 解密 key " + txtAID.Text);
+                    info("利用 AID " + txtAID.Text + "获取 CMA 解密 key ");
                     encKey = Utility.GetEncKey(txtAID.Text);
                     if (encKey.Length != 64) return;
                     info("已获取 CMA 解密 key " + encKey);
@@ -216,11 +292,11 @@ namespace auto_h_encore {
                     incrementProgress();
                     info("auto h-encore 操作完成!!");
                 } catch (DirectoryNotFoundException ex) {
-                    MessageBox.Show("你的 QCMA 目录不见了! 确保你选择了正确的目录,且该目录未被删除!");
+                    MessageBox.Show("找不到你的 QCMA 目录! 确保你选择了正确的目录,且该目录未被删除!");
                     toggleControls(true);
                     return;
                 } catch (UnauthorizedAccessException ex) {
-                    MessageBox.Show("本程序无法写入你的 QCMA 备份目录. 在 QCMA 设置界面修改该目录为你拥有读写权限的目录, 以管理员权限运行, 或者禁用该目录只读权限.");
+                    MessageBox.Show("本程序无法写入你的 QCMA 备份目录. 请在 QCMA 设置界面修改该目录为你拥有读写权限的目录, 或者以管理员权限运行本程序, 或者禁用该目录的只读权限.");
                     toggleControls(true);
                     return;
                 } catch (IOException ex) {
@@ -231,14 +307,15 @@ namespace auto_h_encore {
 
                 Invoke(new Action(() => MessageBox.Show("即将完成 h-encore 安装:\r\n"
                     + "1. 右键点击 QCMA 任务栏图标并选择刷新数据库\r\n"
-                    + "2. 用USB连接PC和PSV\r\n"
-                    + "3. 打开 内容管理 程序并点击 复制内容\r\n"
-                    + "    如果Vita提示你需要更新系统, 关闭 Wifi 并重启 Vita\r\n"
-                    + "4. 在 内容管理 程序, 选择 PC -> PS Vita System\r\n"
+                    + "2. 用USB连接 你的电脑 和 PSV\r\n"
+                    + "3. 打开Vita端 内容管理 程序并点击 复制内容\r\n"
+                    + "     如果 Vita 提示你需要更新系统, 关闭 Wifi 后重启 Vita\r\n"
+                    + "4. 在 内容管理 程序界面, 选择 PC -> PS Vita System\r\n"
                     + "5. 选择 应用程序\r\n"
                     + "6. 选择 PS Vita\r\n"
                     + "7. 选择 h-encore 并点击 复制\r\n"
                     + "8. 启动 h-encore 程序\r\n"
+                    + "     如果初次启动崩溃, 尝试重启Vita后再运行该程序\r\n\r\n"
                     + "完成!")));
 
                 toggleControls(true);
@@ -268,9 +345,9 @@ namespace auto_h_encore {
             VerifyUserInfo();
         }
 
-        private void lblQCMA_Click(object sender, EventArgs e)
-        {
-
+        private void btnImport_Click(object sender, EventArgs e) {
+            FormFiles frm = new FormFiles();
+            frm.ShowDialog();
         }
     }
 }
